+++ Scripts/Engines/BulkOrders/Bods/SmallBOD.cs
@@ -1,7 +1,7 @@
-using System;
-using System.Collections.Generic;
 using Server.Items;
 using Server.Mobiles;
+using System;
+using System.Collections.Generic;
 
 namespace Server.Engines.BulkOrders
 {
@@ -234,7 +234,7 @@ namespace Server.Engines.BulkOrders
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 
 			writer.Write(m_AmountCur);
 			writer.Write(m_AmountMax);
