+++ Scripts/Engines/BulkOrders/Bods/Tailor/SmallTailorBOD.cs
@@ -1,6 +1,6 @@
+using Server.Engines.Craft;
 using System;
 using System.Collections.Generic;
-using Server.Engines.Craft;
 
 namespace Server.Engines.BulkOrders
 {
@@ -221,7 +221,7 @@ namespace Server.Engines.BulkOrders
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
