+++ Scripts/Engines/MLQuests/Definitions/Heritage.cs
@@ -1,8 +1,8 @@
-using System;
 using Server.Engines.MLQuests.Objectives;
 using Server.Engines.MLQuests.Rewards;
 using Server.Items;
 using Server.Mobiles;
+using System;
 
 namespace Server.Engines.MLQuests.Definitions
 {
@@ -263,7 +263,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -295,7 +295,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -330,7 +330,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -408,7 +408,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -437,7 +437,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -469,7 +469,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -501,7 +501,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -685,7 +685,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -727,7 +727,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -780,7 +780,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
