+++ Scripts/Engines/MLQuests/Definitions/Spellweaving.cs
@@ -1,8 +1,8 @@
-using System;
 using Server.Engines.MLQuests.Objectives;
 using Server.Engines.MLQuests.Rewards;
 using Server.Items;
 using Server.Mobiles;
+using System;
 
 namespace Server.Engines.MLQuests.Definitions
 {
@@ -491,7 +491,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -557,7 +557,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -617,7 +617,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
@@ -674,7 +674,7 @@ namespace Server.Engines.MLQuests.Definitions
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
