+++ Scripts/Engines/MLQuests/Items/ABauble.cs
@@ -1,31 +1,31 @@
 namespace Server.Items
 {
-    public class ABauble : BaseItem
-    {
-        public override int LabelNumber { get { return 1073137; } } // A bauble
+	public class ABauble : BaseItem
+	{
+		public override int LabelNumber { get { return 1073137; } } // A bauble
 
-        [Constructable]
-        public ABauble() : base(0x23B)
-        {
-            LootType = LootType.Blessed;
-        }
+		[Constructable]
+		public ABauble() : base(0x23B)
+		{
+			LootType = LootType.Blessed;
+		}
 
-        public ABauble(Serial serial) : base(serial)
-        {
-        }
+		public ABauble(Serial serial) : base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
 
-            writer.Write((int)0); // Version
-        }
+			writer.Write(0); // Version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
 
-            int version = reader.ReadInt();
-        }
-    }
+			int version = reader.ReadInt();
+		}
+	}
 }
