+++ Scripts/Engines/MLQuests/Items/FriendsOfTheLibraryApplication.cs
@@ -1,39 +1,39 @@
 namespace Server.Items
 {
-    public class FriendsOfTheLibraryApplication : BaseItem
-    {
-        public override int LabelNumber { get { return 1073131; } } // Friends of the Library Application
+	public class FriendsOfTheLibraryApplication : BaseItem
+	{
+		public override int LabelNumber { get { return 1073131; } } // Friends of the Library Application
 
-        public override bool Nontransferable { get { return true; } }
+		public override bool Nontransferable { get { return true; } }
 
-        public override void AddNameProperties(ObjectPropertyList list)
-        {
-            base.AddNameProperties(list);
-            AddQuestItemProperty(list);
-        }
+		public override void AddNameProperties(ObjectPropertyList list)
+		{
+			base.AddNameProperties(list);
+			AddQuestItemProperty(list);
+		}
 
-        [Constructable]
-        public FriendsOfTheLibraryApplication() : base(0xEC0)
-        {
-            LootType = LootType.Blessed;
-        }
+		[Constructable]
+		public FriendsOfTheLibraryApplication() : base(0xEC0)
+		{
+			LootType = LootType.Blessed;
+		}
 
-        public FriendsOfTheLibraryApplication(Serial serial) : base(serial)
-        {
-        }
+		public FriendsOfTheLibraryApplication(Serial serial) : base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
 
-            writer.Write((int)0); // Version
-        }
+			writer.Write(0); // Version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
 
-            int version = reader.ReadInt();
-        }
-    }
+			int version = reader.ReadInt();
+		}
+	}
 }
