+++ Scripts/Engines/MLQuests/Items/SealingWaxOrderAddressedToPetrus.cs
@@ -1,39 +1,39 @@
 namespace Server.Items
 {
-    public class SealingWaxOrderAddressedToPetrus : BaseItem
-    {
-        public override int LabelNumber { get { return 1073132; } } // Sealing Wax Order addressed to Petrus
+	public class SealingWaxOrderAddressedToPetrus : BaseItem
+	{
+		public override int LabelNumber { get { return 1073132; } } // Sealing Wax Order addressed to Petrus
 
-        public override bool Nontransferable { get { return true; } }
+		public override bool Nontransferable { get { return true; } }
 
-        public override void AddNameProperties(ObjectPropertyList list)
-        {
-            base.AddNameProperties(list);
-            AddQuestItemProperty(list);
-        }
+		public override void AddNameProperties(ObjectPropertyList list)
+		{
+			base.AddNameProperties(list);
+			AddQuestItemProperty(list);
+		}
 
-        [Constructable]
-        public SealingWaxOrderAddressedToPetrus() : base(0xEBF)
-        {
-            LootType = LootType.Blessed;
-        }
+		[Constructable]
+		public SealingWaxOrderAddressedToPetrus() : base(0xEBF)
+		{
+			LootType = LootType.Blessed;
+		}
 
-        public SealingWaxOrderAddressedToPetrus(Serial serial) : base(serial)
-        {
-        }
+		public SealingWaxOrderAddressedToPetrus(Serial serial) : base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
 
-            writer.Write((int)0); // Version
-        }
+			writer.Write(0); // Version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
 
-            int version = reader.ReadInt();
-        }
-    }
+			int version = reader.ReadInt();
+		}
+	}
 }
