+++ Scripts/Engines/MLQuests/Items/SeveredElfEars.cs
@@ -1,36 +1,36 @@
 namespace Server.Items
 {
-    [Flipable(0x312D, 0x312E)]
-    public class SeveredElfEars : BaseItem
-    {
-        [Constructable]
-        public SeveredElfEars() : this(1)
-        {
-        }
+	[Flipable(0x312D, 0x312E)]
+	public class SeveredElfEars : BaseItem
+	{
+		[Constructable]
+		public SeveredElfEars() : this(1)
+		{
+		}
 
-        [Constructable]
-        public SeveredElfEars(int amount) : base(Utility.RandomList(0x312D, 0x312E))
-        {
-            Stackable = true;
-            Amount = amount;
-        }
+		[Constructable]
+		public SeveredElfEars(int amount) : base(Utility.RandomList(0x312D, 0x312E))
+		{
+			Stackable = true;
+			Amount = amount;
+		}
 
-        public SeveredElfEars(Serial serial) : base(serial)
-        {
-        }
+		public SeveredElfEars(Serial serial) : base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
 
-            writer.Write((int)0); // Version
-        }
+			writer.Write(0); // Version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
 
-            int version = reader.ReadInt();
-        }
-    }
+			int version = reader.ReadInt();
+		}
+	}
 }
